Altere o tipo de pessoa de int para Long

+++ src/main/java/com/marhashoft/requerimentoapi/repository/PessoaRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface PessoaRepository extends JpaRepository<Pessoa, Long> {
 
-    List<Pessoa> findByAtivoTrueAndTipoPessoaIdOrderByNome(int tipoPessoaId);
+    List<Pessoa> findByAtivoTrueAndTipoPessoaIdOrderByNome(Long tipoPessoaId);
 
     Long countByTipoPessoaId(Long tipoPessoaId);
 }
